Move dispatch related APIs to Node and fix issues
